Change kernel commandline for target kernel under test
Add "printk.time=y quiet loglevel=5" to the command line of the Linux
kernel under test in /etc/rc.local. Timestamps are enabled to be able
to separate Linux kernel output from userspace messages in an easy
manner. The parameters "quiet" and "loglevel=5" are used to reduce
Linux kernel output to warnings and errors.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>