rc.local: Remove root parameter detection heuristic
The linux kernel root parameter detection heuristic in /etc/rc.local
tent to be error prone. The heuristic was introduced due to the fact,
that the kernel command line and especially the root parameter need to
be changed to fit the kernel under test. E.L.B.E. use LABEL to specify
the linux kernel root parameter. LABEL is also used in /etc/fstab and
do not differ after the change to the linux kernel root parameter any
more.

Remove the root parameter detection heuristic in /etc/rc.local and use
LABEL to specify root instead.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Holger Dengler <dengler@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>